Update style.css
Made the proper percentage chart as shown in the skillset portion of class bar and at heading h3. Actually, before it was showing like 65% but at that moment it was not showing the bars width upto 65. Hence, I have just made the every child of it showing the bar according to it's perentage.